+++ src/managers/DomUIManager.ts
@@ -400,6 +400,11 @@ export class DomUIManager implements IUIManager {
       menu.push({
         label,
         callback: () => {
+          // 节点坐标减去画布偏移坐标量
+          if (node.viewer){
+            position.x -= node.viewer.events.viewPosition.x
+            position.y -= node.viewer.events.viewPosition.y
+          }
           this.events.dispatch(GraphAction.AddNode, type, position);
         }
       })