+++ src/renders/DomRender/DomNodeRender.ts
@@ -106,6 +106,14 @@ export class DomNodeRender implements INodeRender {
         color: this.node.options['nodeFontColor'] ?? config.style.NodeFontColor,
       })
 
+    // 节点拖拽宽高还原
+    if (node.getOption('clientWidth')) {
+      root.setWidth(node.getOption('clientWidth'), 'min')
+    }
+    if (node.getOption('clientHeight')) {
+      root.setHeight(node.getOption('clientHeight'), 'min')
+    }
+
     title.setWidth('100%');
     title.setHeight(titleHeight, 'min');
     title.setStyle({
@@ -406,7 +414,7 @@ export class DomNodeRender implements INodeRender {
 
     title.innerText(this.node.label);
 
-    if (DomNodeRender.showNodeIndex) {
+    if (this.events.viewer.option.nodeIndexShow && DomNodeRender.showNodeIndex) {
       // @mark 给title加上左侧的index显示
       let indexBox = new NativeDiv()
       indexBox.innerText(this.node.index.toString() ?? "\\")