+++ src/shared/UI/widgets/WidgetsManager.ts
@@ -3,7 +3,7 @@
  * @Description: 微件管理器
  * @FilePath: /bimcc-graph/src/shared/UI/widgets/WidgetsManager.ts
  * @Date: 2023-10-07 15:36:41
- * @LastEditTime: 2023-10-12 11:16:23
+ * @LastEditTime: 2023-11-08 15:22:35
  * @Author: lisushuang
  */