+++ changelog.md
@@ -1,3 +1,15 @@
+【v1.3.0】
+
+新增无图直接运行的选项
+
+实例化GraphViewer的时候不传入dom即可
+
+同时记得设置 viewer.graph.setStepTime(0)
+
+开始执行：viewer.grpah.start()
+
+停止执行：viewer.graph.stop()
+
 【v1.0.0】
 
 项目创建